+++ R/mutate_location.R
@@ -2,16 +2,24 @@
 #'
 #' Convert AMLR location strings to accepted values
 #'
-#' @param x data.frame, must contain character column location
+#' @param x data.frame, must contain character column 'location'
 #'
 #' @details
 #' AMLR (pinniped) data is frequently messy, including having non-standard
 #' location information (location names). This function is a 'repository'
 #' of approved conversions to standardize location names
 #' in an ***REMOVED*** dataset.
+#' This function was built for the tag resight reimport effort,
+#' but is applicable to other location renaming as well
+#'
+#' Input data are processed as followed before being matched:
+#'   - '.' characters are removed (i.e., replaced with '')
+#'   - Non-ASCII text is escaped using
+#'     \code{\link[stringi:stri_escape_unicode]{stri_escape_unicode}}
 #'
 #' @return
-#' A data frame, with the same columns and types.
+#' A data frame, with an updated 'locations' column
+#' but otherwise the same columns and types.
 #'
 #' @examples
 #' x <- data.frame(id = 1:3, location = c("P Bahamonde", "Bahamonde", "Chung"))
@@ -25,13 +33,18 @@ mutate_location <- function(x) {
     stop("The data frame must have a column location of type character")
 
   x %>%
-    mutate(loc_lower = str_replace_all(tolower(location), "[.]", ""),
+    mutate(loc_clean = str_replace_all(stri_escape_unicode(location), "[.]", ""),
+           loc_lower = tolower(loc_clean),
+           # https://stackoverflow.com/questions/22223431
+           # stringi::stri_escape_unicode(c("á", "í", "ú"))
            location =
              case_when(
-               loc_lower %in% tolower(c("AH pt")) ~ "Odontoceto, Punta",
+               loc_lower %in% tolower(c("AH pt")) ~ "Punta Odontoceto",
                loc_lower %in% tolower(c("Playa Alcazar", "Alcazor", "Alcazar flats", "Planicie Alcazar", "PAlcazar", "P Alcazar")) ~ "Alcazar",
-               loc_lower %in% tolower(c("Angosta", "Pl Angosta", "P Angosta")) ~ "Angosta, Playa",
-               loc_lower %in% tolower(c("Antarctica", "Playa Antarctica", "Pl Antarctica", "Antártico", "P Antarctico", " P Antarctica", "P Antarctica", "Antartica", "Antarctico")) ~ "Antartico, Playa",
+               loc_lower %in% tolower(c("Angosta", "Pl Angosta", "P Angosta")) ~ "Angosta",
+               # loc_lower %in% tolower(c("Antarctica", "Playa Antarctica", "Pl Antarctica", "Antártico", "P Antarctico", " P Antarctica", "P Antarctica", "Antartica", "Antarctico")) ~ "Antartico",
+               loc_lower %in% tolower(c("Antarctica", "Playa Antarctica", "Pl Antarctica", paste0("Ant", "\\u00e1", "rtico"), "P Antarctico", " P Antarctica", "P Antarctica", "Antartica", "Antarctico")) ~ "Antartico",
+
                loc_lower %in% tolower(c(" Amarya", "Amarya", "behind Elmira", "Elmira", "Aymara F")) ~ "Aymara",
                loc_lower %in% tolower(c("P Bahamonde", "Bahamunde", "B'monde/La Cav", "B Mon", "Bajamonde", "Baha Monde", "B'monde")) ~ "Bahamonde",
 
@@ -77,10 +90,10 @@ mutate_location <- function(x) {
                loc_lower %in% tolower(c("Hue Far E")) ~ "Hue Far East",
                loc_lower %in% tolower(c("Hue S")) ~ "Hue South",
 
-               loc_lower %in% tolower(c("Playa Lobero", "Lobero")) ~ "del Lobero, Playa",
-               loc_lower %in% tolower(c("Plastico")) ~ "El Plastico, Playa",
+               loc_lower %in% tolower(c("Playa Lobero", "Lobero")) ~ "del Lobero",
+               loc_lower %in% tolower(c("Plastico")) ~ "El Plastico",
                loc_lower %in% tolower(c("Monument")) ~ "Gaviota",
-               loc_lower %in% tolower(c("Punta Haydee")) ~ "Haydee, Punta",
+               loc_lower %in% tolower(c("Punta Haydee")) ~ "Punta Haydee",
                loc_lower %in% tolower(c("taco", "Cerro Taco", "Hill 2")) ~ "Huillin",
                loc_lower %in% tolower(c("Broken Arrow")) ~ "Cerro Parmenio",
                loc_lower %in% tolower(c("Abismo", "ALC abismo")) ~ "Cerro El Abismo",
@@ -89,7 +102,8 @@ mutate_location <- function(x) {
                loc_lower %in% tolower(c("lago oculto")) ~ "Lago Oculto",
                loc_lower %in% tolower(c("La Caverna", "Lacaverna", "La Cavarna", "Caverna")) ~ "La Caverna",
 
-               loc_lower %in% tolower(c("Lobería", "Lobe ria", "loberia", "loberia central", "Loberia C", "Pl Lobería", "Lob C")) ~ "Loberia",
+               # loc_lower %in% tolower(c(paste0("Lober", "\\u00ed", "a"), "Lobe ria", "loberia", "loberia central", "Loberia C", "Pl Lobería", "Lob C")) ~ "Loberia",
+               loc_lower %in% tolower(c(paste0("Lober", "\\u00ed", "a"), "Lobe ria", "loberia", "loberia central", "Loberia C", paste0("Pl Lober", "\\u00ed", "a"), "Lob C")) ~ "Loberia",
                loc_lower %in% tolower(c("Loberia N", "Loberia-N", "Lobo N")) ~ "Loberia North",
                loc_lower %in% tolower(c("Loberia S", "Loberia S", "Loveria S", "Lob S")) ~ "Loberia South",
                loc_lower %in% tolower(c("Lob flats", "Loberia F", "Loberia flats", "Loberia flats S", "lobe ria valley", "Loberia Fl")) ~ "Loberia Flats",
@@ -103,7 +117,9 @@ mutate_location <- function(x) {
                loc_lower %in% tolower(c("Mad-r", "Mad R", "Mad Ridge", "Maderas Rdg")) ~ "Maderas Ridge",
                loc_lower %in% tolower(c("Mad Saddle")) ~ "Maderas Saddle",
 
-               loc_lower %in% tolower(c("Módulo", "PMod", "Modelo", "landing beach", "Mod", "modulo", "Pl El Modulo", "El Módulo", "Playa Modulo", "Above int'l highway", "International highway", "PMod", "Highway",
+               # loc_lower %in% tolower(c("Módulo", "PMod", "Modelo", "landing beach", "Mod", "modulo", "Pl El Modulo", "El Módulo", "Playa Modulo", "Above int'l highway", "International highway", "PMod", "Highway",
+               loc_lower %in% tolower(c(paste0("M", "\\u00f3", "dulo"), "PMod", "Modelo", "landing beach", "Mod", "modulo", "Pl El Modulo", paste0("El M", "\\u00f3", "dulo"),
+                                        "Playa Modulo", "Above int'l highway", "International highway", "PMod", "Highway",
                                         "Mod C", "Mod C flat", "Mod C high", "Mod E", "Mod E high", "Mod flat", "Mod H", "Mod low", "Mod W", "Mod W flats", "Mod W high",
                                         "Modulo", "Modulo Ridge", "P el Modulo", "Mod N", "Modulo C")) ~ "Modulo",
                loc_lower %in% tolower(c("P Modulo N")) ~ "Modulo North",
@@ -120,28 +136,31 @@ mutate_location <- function(x) {
                loc_lower %in% tolower(c("P Media Luna", "media luna", "Media Luna flats", "Medi Luna", "Meda Luna", "Meda Luna S", "Media L", "S Pl Media Luna", "M Luna", "MLuna", "MML", "ML", "towards Media Luna")) ~ "Media Luna",
 
                loc_lower %in% tolower(c("Ivaldo", "Pl N Baldo", "Nivaldo", "Nibaldo Bahamondes", "P Nibaldo")) ~ "Nibaldo",
-               loc_lower %in% tolower(c("PNegra", "PNegra east", "P Negra")) ~ "Negra, Punta",
+               loc_lower %in% tolower(c("PNegra", "PNegra east", "P Negra")) ~ "Punta Negra",
                loc_lower %in% tolower(c("P Elevado", "Paso elevado")) ~ "Paso Elevado",
-               loc_lower %in% tolower(c("Papúa", "Playa Papua", "P Papua")) ~ "Papua",
-               loc_lower %in% tolower(c("Pl Paulina", "West Coast central", "WST")) ~ "Paulina, Playa",
+               # loc_lower %in% tolower(c("Papúa", "Playa Papua", "P Papua")) ~ "Papua",
+               loc_lower %in% tolower(c(paste0("Pap", "\\u00fa", "a"), "Playa Papua", "P Papua")) ~ "Papua",
+               loc_lower %in% tolower(c("Pl Paulina", "West Coast central", "WST")) ~ "Paulina",
                loc_lower %in% tolower(c("Peng Col 12", "Col 14", "Colony 23", "Colony 20", "Colony 10", "Col 23")) ~ "Penguin Colonies",
                loc_lower %in% tolower(c("Pinochet", "P Pinochet de la Barra")) ~ "Pinochet de la Barra",
-               loc_lower %in% tolower(c("Punta Pobleta", "Punta Poblete", "Poblete")) ~ "Poblete, Punta",
-               loc_lower %in% tolower(c("Pocitas", "P Pocitas")) ~ "Pocitas, Playa",
-               loc_lower %in% tolower(c("Sciappacasse")) ~ "Schiappacasse, Playa",
+               loc_lower %in% tolower(c("Punta Pobleta", "Punta Poblete", "Poblete")) ~ "Punta Poblete",
+               loc_lower %in% tolower(c("Pocitas", "P Pocitas")) ~ "Pocitas",
+               loc_lower %in% tolower(c("Sciappacasse")) ~ "Schiappacasse",
                loc_lower %in% tolower(c("Punta Ventana")) ~ "Ventana",
                loc_lower %in% tolower(c("Flats Toqui", "Tokey")) ~ "Toqui",
-               loc_lower %in% tolower(c("Roquerío", "P Roquerio")) ~ "Roquerio",
-               loc_lower %in% tolower(c("Remanso")) ~ "El Remanso, Playa",
-               loc_lower %in% tolower(c("San Telmo", "San Telmo Norte", "P San Telmo", "Punta San Telmo")) ~ "San Telmo, Punta",
+               # loc_lower %in% tolower(c("Roquerío", "P Roquerio")) ~ "Roquerio",
+               loc_lower %in% tolower(c(paste0("Roquer", "\\u00ed", "o"), "P Roquerio")) ~ "Roquerio",
+               loc_lower %in% tolower(c("Remanso")) ~ "El Remanso",
+               loc_lower %in% tolower(c("San Telmo", "San Telmo Norte", "P San Telmo", "Punta San Telmo")) ~ "PuntaSan Telmo",
                loc_lower %in% tolower(c("Valley Corto")) ~ "Valle Corto",
-               loc_lower %in% tolower(c("P Yamana", "Yámana", "Yamana", "Yamana Flats", "Pl Yamana", "Playa Yamana")) ~ "Yamana, Playa",
-               loc_lower %in% tolower(c("P Yuseff", "ML/P Yusseff", "P Yusseff", "Punta Yuseff")) ~ "Yuseff, Punta",
+               # loc_lower %in% tolower(c("P Yamana", "Yámana", "Yamana", "Yamana Flats", "Pl Yamana", "Playa Yamana")) ~ "Yamana",
+               loc_lower %in% tolower(c("P Yamana", paste0("Y", "\\u00e1", "mana"), "Yamana", "Yamana Flats", "Pl Yamana", "Playa Yamana")) ~ "Yamana",
+               loc_lower %in% tolower(c("P Yuseff", "ML/P Yusseff", "P Yusseff", "Punta Yuseff")) ~ "Punta Yuseff",
 
-               # TODO: "P Golondrina to P del Lobero", "P Delphin to P La Caverna", "P Paulina to P Aranda"
+               # "P Golondrina to P del Lobero", "P Delphin to P La Caverna", "P Paulina to P Aranda"
                # loc_lower %in% tolower(c("WST")) ~ "Unknown",
 
-               TRUE ~ location
+               .default = location
              )) %>%
-    select(-loc_lower)
+    select(-c(loc_lower, loc_clean))
 }